+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - pull validators list from target blockchain
 - do not overshoot heartbeat
+- cache required number of signatures
 
 ## [7.10.0] - 2023-06-15
 ### Added